Convert webhook controller to typescript
