+++ src/app/modules/webhook/webhook.controller.ts
@@ -1,16 +1,7 @@
 import { NextFunction, Request, Response } from 'express'
 
-import formsgSdk from '../../../config/formsg-sdk'
-
-import { WebhookValidationError } from './webhook.errors'
-import {
-  handleWebhookFailure,
-  handleWebhookSuccess,
-  logWebhookFailure,
-  postWebhook,
-} from './webhook.service'
+import { pushData } from './webhook.service'
 import { WebhookRequestLocals } from './webhook.types'
-import { validateWebhookUrl } from './webhook.utils'
 
 /**
  * POST submission to a specified URL.  Only works for encrypted submissions.
@@ -28,52 +19,10 @@ export function post(
   next: NextFunction,
 ) {
   const { form, submission } = req
-  if (form.webhook.url) {
-    const webhookUrl = form.webhook.url
-    const now = Date.now()
-    const submissionWebhookView = submission.getWebhookView()
-
-    // Log and return, this should not happen.
-    if (!submissionWebhookView) {
-      logWebhookFailure(
-        new WebhookValidationError('submissionWebhookView was null'),
-        {
-          webhookUrl,
-          submissionWebhookView,
-          now,
-        },
-      )
-      return next()
-    }
-
-    const { submissionId, formId } = submissionWebhookView.data
-
-    const signature = formsgSdk.webhooks.generateSignature({
-      uri: form.webhook.url,
-      submissionId,
-      formId,
-      epoch: now,
-    }) as string
-
-    const webhookParams = {
-      webhookUrl,
-      submissionWebhookView,
-      submissionId,
-      formId,
-      now,
-      signature,
-    }
-
-    // Use promises instead of await to prevent the user from having to await on the
-    // webhook before they receive acknowledgement that their submission was successful.
-    validateWebhookUrl(webhookParams.webhookUrl)
-      .then(() => postWebhook(webhookParams))
-      .then((response) => {
-        handleWebhookSuccess(response, webhookParams)
-      })
-      .catch((error) => {
-        handleWebhookFailure(error, webhookParams)
-      })
+  const webhookUrl = form.webhook.url
+  const submissionWebhookView = submission.getWebhookView()
+  if (webhookUrl) {
+    pushData(webhookUrl, submissionWebhookView)
   }
   return next()
 }
