+++ src/app/modules/webhook/webhook.service.ts
@@ -10,11 +10,13 @@ import {
   IFormSchema,
   ISubmissionSchema,
   IWebhookResponse,
+  WebhookView,
 } from '../../../types'
 import { getEncryptSubmissionModel } from '../../models/submission.server.model'
 
 import { WebhookValidationError } from './webhook.errors'
 import { WebhookParams } from './webhook.types'
+import { validateWebhookUrl } from './webhook.utils'
 
 const logger = createLoggerWithLabel(module)
 const EncryptSubmission = getEncryptSubmissionModel(mongoose)
@@ -30,7 +32,7 @@ const EncryptSubmission = getEncryptSubmissionModel(mongoose)
  * @param {string} webhookParams.now Epoch for POST header
  * @param {string} webhookParams.signature Signature generated by FormSG SDK
  */
-export const handleWebhookFailure = async (
+const handleWebhookFailure = async (
   error: Error | AxiosError,
   webhookParams: WebhookParams,
 ): Promise<any> => {
@@ -53,7 +55,7 @@ export const handleWebhookFailure = async (
  * @param {string} webhookParams.now Epoch for POST header
  * @param {string} webhookParams.signature Signature generated by FormSG SDK
  */
-export const handleWebhookSuccess = async (
+const handleWebhookSuccess = async (
   response: AxiosResponse,
   webhookParams: WebhookParams,
 ): Promise<any> => {
@@ -76,7 +78,7 @@ export const handleWebhookSuccess = async (
  * @param {string} now Epoch for POST header
  * @param {string} signature Signature generated by FormSG SDK
  */
-export const postWebhook = ({
+const postWebhook = ({
   webhookUrl,
   submissionWebhookView,
   submissionId,
@@ -119,7 +121,7 @@ const logWebhookSuccess = (
 }
 
 // Logging for webhook failure
-export const logWebhookFailure = (
+const logWebhookFailure = (
   error: Error | AxiosError,
   { webhookUrl, submissionId, formId, now, signature }: Partial<WebhookParams>,
 ): void => {
@@ -182,7 +184,7 @@ const updateSubmissionsDb = async (
 // Formats webhook success info into an object to update Submissions collection
 const getSuccessDbUpdate = (
   response: AxiosResponse,
-  { webhookUrl, signature }: WebhookParams,
+  { webhookUrl, signature }: Pick<WebhookParams, 'webhookUrl' | 'signature'>,
 ): IWebhookResponse => {
   return { webhookUrl, signature, ...getFormattedResponse(response) }
 }
@@ -190,7 +192,7 @@ const getSuccessDbUpdate = (
 // Formats webhook failure info into an object to update Submissions collection
 const getFailureDbUpdate = (
   error: Error | AxiosError,
-  { webhookUrl, signature }: WebhookParams,
+  { webhookUrl, signature }: Pick<WebhookParams, 'webhookUrl' | 'signature'>,
 ): IWebhookResponse => {
   const errorMessage = get(error, 'message')
   let update: IWebhookResponse = {
@@ -217,4 +219,52 @@ const getFormattedResponse = (
       data: stringifySafe(get(response, 'data')),
     },
   }
+}
+
+export const pushData = (
+  webhookUrl: WebhookParams['webhookUrl'],
+  submissionWebhookView: WebhookView,
+): Promise<any> => {
+  const now = Date.now()
+  // Log and return, this should not happen.
+  if (!submissionWebhookView) {
+    logWebhookFailure(
+      new WebhookValidationError('submissionWebhookView was null'),
+      {
+        webhookUrl,
+        submissionWebhookView,
+        now,
+      },
+    )
+    return
+  }
+
+  const { submissionId, formId } = submissionWebhookView.data
+
+  const signature = formsgSdk.webhooks.generateSignature({
+    uri: webhookUrl,
+    submissionId,
+    formId,
+    epoch: now,
+  }) as string
+
+  const webhookParams = {
+    webhookUrl,
+    submissionWebhookView,
+    submissionId,
+    formId,
+    now,
+    signature,
+  }
+
+  // Use promises instead of await to prevent the user from having to await on the
+  // webhook before they receive acknowledgement that their submission was successful.
+  return validateWebhookUrl(webhookParams.webhookUrl)
+    .then(() => postWebhook(webhookParams))
+    .then((response) => {
+      handleWebhookSuccess(response, webhookParams)
+    })
+    .catch((error) => {
+      handleWebhookFailure(error, webhookParams)
+    })
 }
